Load fuzzy finder paths in web worker at startup
Previously the paths were first loaded when the fuzzy finder
view was first attached.

Now a web worker is started when the package activates that
sets the paths on the fuzzy finder view the first time it
displays if the worker has completed by the time an event occurs
that displays the view.