:racehorse: Don't create project view until command is run
Previously the project view was created in order to start the load
task which meant the views were created before they were used and slowed
down startup.

Now the task is kicked off outside of the view and the paths are passed
into the view when the task completes and the task is run.

This restores the prior behavior that did this same technique to minimally
impact startup time.