Use scandal to build the project file list
Atom core has scandal, which is _all about_ building a list of files
in a fairly efficient manner, and thinking about VCS issues and
symbolic links and whatnot. Instead of reimplementing pieces of that,
just use it.

This has the nice side-effect of fixing issue #111, because scandal is
better at deciding which files should be ignored.