Make fuzzy finder tests more deterministic
The modified test was relying on the order that two crawlers that
are running in parallel start returning results. This may have probably caused
flakiness and unintended failures.

In order to fix this, I'm forcing the crawling system to run
sequentially, by faking that the system has a single CPU.