Fix excluding VCS ignored paths based on the flag in core settings
The `ignoreVcsIgnores` property was not set in the PathLoader constructor and
when it was used later it didn't hold any value so the option to not ignore
the VCS ignored paths could never be passed to ripgrep. As ignoring is the
default behavior, adding the option to NOT ignore needs to be done when
`core.excludeVcsIgnoredPaths` is `false`.